# This seeming small optimization has a surprisingly big effect.