@Sheikah45 2024-07-14
I also suspect it has such a large impact because getSize has two other boolean checks inside it as well.